// client/src/api/models/postApiCommunitiesBody.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Readdit
 * OpenAPI spec version: v1
 */
import type { CommunityType } from "./communityType";

export type PostApiCommunitiesBody = {
   Name: string;
   Description: string;
   Picture?: Blob;
   Tags?: string[];
   Type: CommunityType;
};